+++ encompass/encompass/urls.py
@@ -78,6 +78,7 @@ urlpatterns = [
     re_path(r"^encompass/about/$", views.about_page),
     re_path(r"^encompass/encapsule_sync/$", views.encapsule_sync_now),
     re_path(r"^encompass/unclassified_hosts/$", views.unclassified_hosts_page),
+    re_path(r"^encompass/spring_cleaning/$", views.spring_cleaning_page),
     re_path(r"^encompass/git_log/$", views.git_log_page),
     re_path(r"^encompass/user_settings/$", views.user_settings),
     re_path(r"^encompass/query_host/$", views.query_host),